Organized extension methods into new files / classes.
