+++ src/Extensions/IFastFileCopy.cs
@@ -5,10 +5,10 @@ using System.Threading.Tasks;
 namespace OwlCore.Storage;
 
 /// <summary>
-/// Provides a fast-path for the <see cref="FolderExtensions.CreateCopyOfAsync"/> extension method.
+/// Provides a fast-path for the <see cref="ModifiableFolderExtensions.CreateCopyAsync"/> extension method.
 /// </summary>
 /// <exception cref="FileNotFoundException">The item was not found in the provided folder.</exception>
-public interface IFastFileCopy<T> : IModifiableFolder
+public interface IFastFileCopy<in T> : IModifiableFolder
     where T : IFile
 {
     /// <summary>
